requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
DAWG2-Python==0.9.0
et_xmlfile==2.0.0
fastapi==0.135.2
h11==0.16.0
idna==3.11
kombu==5.6.2
openpyxl==3.1.5
packaging==26.0
prompt_toolkit==3.0.52
pydantic==2.12.5
pydantic_core==2.41.5
pymorphy3==2.0.6
pymorphy3-dicts-ru==2.4.417150.4580142
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.4.0
setuptools==82.0.1
six==1.17.0
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.42.0
vine==5.1.0
wcwidth==0.6.0
